+++ rust/user-info-fetcher/src/main.rs
@@ -13,7 +13,7 @@ use reqwest::ClientBuilder;
 use serde::{Deserialize, Serialize};
 use snafu::{ResultExt, Snafu};
 use stackable_opa_operator::crd::user_info_fetcher::v1alpha1;
-use stackable_operator::telemetry::Tracing;
+use stackable_operator::{commons::tls_verification::TlsClientDetails, telemetry::Tracing};
 use tokio::net::TcpListener;
 
 mod backend;
@@ -144,6 +144,10 @@ async fn main() -> Result<(), StartupError> {
             client_id: "".to_string(),
             client_secret: "".to_string(),
         },
+        v1alpha1::Backend::Entra(_) => Credentials {
+            client_id: read_config_file(&args.credentials_dir.join("clientId")).await?,
+            client_secret: read_config_file(&args.credentials_dir.join("clientSecret")).await?,
+        },
     });
 
     let mut client_builder = ClientBuilder::new();
@@ -156,6 +160,15 @@ async fn main() -> Result<(), StartupError> {
         client_builder = utils::tls::configure_reqwest(&keycloak.tls, client_builder)
             .await
             .context(ConfigureTlsSnafu)?;
+    } else if let v1alpha1::Backend::Entra(entra) = &config.backend {
+        client_builder = utils::tls::configure_reqwest(
+            &TlsClientDetails {
+                tls: entra.tls.clone(),
+            },
+            client_builder,
+        )
+        .await
+        .context(ConfigureTlsSnafu)?;
     }
     let http = client_builder.build().context(ConstructHttpClientSnafu)?;
 
@@ -253,6 +266,9 @@ enum GetUserInfoError {
     ActiveDirectory {
         source: backend::active_directory::Error,
     },
+
+    #[snafu(display("failed to get user information from Entra"))]
+    Entra { source: backend::entra::Error },
 }
 
 impl http_error::Error for GetUserInfoError {
@@ -267,6 +283,7 @@ impl http_error::Error for GetUserInfoError {
             Self::Keycloak { source } => source.status_code(),
             Self::ExperimentalXfscAas { source } => source.status_code(),
             Self::ActiveDirectory { source } => source.status_code(),
+            Self::Entra { source } => source.status_code(),
         }
     }
 }
@@ -326,6 +343,11 @@ async fn get_user_info(
                         )
                         .await
                         .context(get_user_info_error::ActiveDirectorySnafu)
+                    }
+                    v1alpha1::Backend::Entra(entra) => {
+                        backend::entra::get_user_info(&req, &http, &credentials, entra)
+                            .await
+                            .context(get_user_info_error::EntraSnafu)
                     }
                 }
             })